factor out AndroidModels
