+++ Sources/SkipAppLovin/AndroidModels/MAAdWaterfallInfo.swift
@@ -1,0 +1,19 @@
+//
+//  MAAdWaterfallInfo.swift
+//  skip-applovin
+//
+//  Created by Dan Fabulich on 1/13/26.
+//
+
+#if !SKIP_BRIDGE
+#if SKIP
+import Foundation
+
+// MARK: - MAAdWaterfallInfo
+
+/// Placeholder for MAAdWaterfallInfo - to be implemented later
+public class MAAdWaterfallInfo {
+    // TODO: Implement MAAdWaterfallInfo wrapper
+}
+#endif
+#endif
