+++ Sources/SkipAppLovin/AndroidModels/MANativeAd.swift
@@ -1,0 +1,19 @@
+//
+//  MANativeAd.swift
+//  skip-applovin
+//
+//  Created by Dan Fabulich on 1/13/26.
+//
+
+#if !SKIP_BRIDGE
+#if SKIP
+import Foundation
+
+// MARK: - MANativeAd
+
+/// Placeholder for MANativeAd - to be implemented later
+public class MANativeAd {
+    // TODO: Implement MANativeAd wrapper
+}
+#endif
+#endif
